Support LinkMode::Alias.
